CrosshairWindow refactored to Singleton, new methods
The `CrosshairSelection` class has been updated to make it easier to manage `CrosshairWindow` instances. Instead of opening multiple windows, a singleton instance of `CrosshairWindow` is now used.

The `Image_MouseLeftButtonUp` method has been changed to use the new singleton instance and manage the image size and rotation controls directly in the main window.

The `RemoveItem`, `MoveItemUp` and `MoveItemDown` methods have been added to make it easier to manage the images in the `CrosshairWindow`.

Updated the `ImageSizeSlider_ValueChanged` method to change the image size in the `CrosshairWindow`.

Updated `RotateImage` method to change image rotation in `CrosshairWindow`.

The `CrosshairWindow` class has been refactored into a singleton to ensure that only one instance of the window exists.

New methods `AddItem`, `RemoveItem`, `MoveItemUp` and `MoveItemDown` have been added to the `CrosshairWindow` class to make it easier to manage the images in the window.

The XAML file `CrosshairWindow.xaml` has been updated to add a named `Grid` element (`MainGrid`) that is used to manage the images.

The `SettingsPage.xaml` file has been updated to add a new "Set hide key" button.

A new method `SetHideKey_Click` has been added to the `SettingsPage.xaml.cs` file to respond to clicks on the new button.